requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.0
crispy-bootstrap5==2024.2
Django==5.0.3
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-multiupload==0.6.1
django-pymemcache==1.0.0
djangorestframework==3.15.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
packaging==24.0
pip-review==1.3.0
psycopg==3.1.18
pymemcache==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
whitenoise==6.6.0
